tidy up and restructure to more match original
