--- .npmrc
@@ -1,1 +1,0 @@
-prefix=/home/kevin/.npm-global
