--- server/authoritative_server/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title>Authoritative Headless Server</title>
-  </head>
-  <body>
-    <!-- load Phaser locally so jsdom doesn't hit a CDN -->
-    <script src="./phaser.min.js"></script>
-
-    <!-- load the authoritative server game loop -->
-    <script src="./js/game.js"></script>
-  </body>
-</html>
